+++ core/src/main/kotlin/sift/core/api/TypeAlias.kt
@@ -1,6 +1,7 @@
 package sift.core.api
 
 typealias IterClasses    = Iterable<Element.Class>
+typealias IterSignatures = Iterable<Element.Signature>
 typealias IterMethods    = Iterable<Element.Method>
 typealias IterFields     = Iterable<Element.Field>
 typealias IterParameters = Iterable<Element.Parameter>