+++ core/src/test/kotlin/sift/core/api/testdata/set1/ArrayListOfPayload.kt
@@ -1,0 +1,4 @@
+package sift.core.api.testdata.set1
+
+class ArrayListOfPayload() : HashMap<String, Payload>()
+