+++ core/src/test/kotlin/sift/core/api/testdata/set1/DictStringT.kt
@@ -1,0 +1,7 @@
+package sift.core.api.testdata.set1
+
+abstract class DictStringT<T> : HashMap<String, T>() {
+
+    val listOfPayload = listOf<Payload>()
+    val listOfT = listOf<T>()
+}